StopWords.jl Documentation

StopWords.jl Documentation

Adding StopWords.jl

Documentation

Index

Adding StopWords.jl

julia> using Pkg
julia> Pkg.add("StopWords")    Updating registry at `~/.julia/registries/General.toml`
   Resolving package versions...
   Installed StopWords ─ v1.0.0
    Updating `~/work/StopWords.jl/StopWords.jl/docs/Project.toml`
  [35248bf2] ~ StopWords v1.0.1 `~/work/StopWords.jl/StopWords.jl` ⇒ v1.0.0
    Updating `~/work/StopWords.jl/StopWords.jl/docs/Manifest.toml`
  [35248bf2] ~ StopWords v1.0.1 `~/work/StopWords.jl/StopWords.jl` ⇒ v1.0.0
Precompiling project...
  ✓ StopWords
  1 dependency successfully precompiled in 1 seconds. 24 already precompiled.
  1 dependency precompiled but a different version is currently loaded. Restart julia to access the new version

Documentation

StopWords.stopwords — Constant

stopwords is a lazy dictionary of stop words for multiple languages. You can access the stop words for a given language using the language's name or ISO 639 code. For example, to get the stop words for English, you can use stopwords["eng"], stopwords["en"], or stopwords["English"]. And to get the stop words for multiple languages, you can use a list, e.g. stopwords[["en", "fr"]].

Base.haskey — Method

haskey(sw::StopWordsDict, lang::AbstractString) -> Bool
haskey(sw::StopWordsDict, langs) -> Bool

Check if lang is a supported language. You can use either the language's name or its ISO 639 code. For example, haskey(stopwords, "en"), haskey(stopwords, "eng"), and haskey(stopwords, "English") all return true. And haskey(stopwords, ["en", "fr"]) returns true if both English and French are supported.

StopWords.supported_languages — Function

supported_languages() -> Set{String}

Return a set containing all the languages (ISO 639-3 codes) that are supported by this package.

Index

StopWords.stopwords

Base.haskey

StopWords.supported_languages