StopWords.jl Documentation

StopWords.jl Documentation

Adding StopWords.jl

Documentation

Index

Adding StopWords.jl

julia> using Pkg
julia> Pkg.add("StopWords")    Updating registry at `~/.julia/registries/General.toml`
   Resolving package versions...
ERROR: expected package `StopWords [35248bf2]` to be registered

Documentation

StopWords.stopwords — Constant

stopwords is a lazy dictionary of stop words for multiple languages. You can access the stop words for a given language using the language's name or ISO 639 code. For example, to get the stop words for English, you can use stopwords["eng"], stopwords["en"], or stopwords["English"].

Base.haskey — Method

haskey(sw::StopWordsDict, lang::String) -> Bool

Check if lang is a supported language. You can use either the language's name or its ISO 639 code. For example, haskey(stopwords, "en"), haskey(stopwords, "eng"), and haskey(stopwords, "English") all return true.

StopWords.supported_languages — Function

supported_languages() -> Vector{String}

Return a vector containing all the languages (ISO 639-3 codes) that are supported by this package.

Index

StopWords.stopwords

Base.haskey

StopWords.supported_languages